Plugin setup
lein-cooper is now a plugin instead of a runnable main class and it has
been refactored to ensure both stderr and stdout are piped to the screen
and dealt with accordingly.

There are problems though. If a command spawns a child process these
aren't killed whrn the progrma closes which is a total ballex.